skip completed jobs
added behaviour to job submission function so any bash scripts that have an associated completed fit will not be resubmitted, which would theoretically allow for a script that runs the analysis_script to be rerun in the event of a timeout without having to worry about wasting cpu time

+++ utils/analysis.py
@@ -328,12 +328,16 @@ def submit_slurm_job(job_path, delete=False, **kwargs):
     Returns:
     - None
     '''
-    
-    submission_cmd = f'sbatch {job_path}'
-    subp = subprocess.run(submission_cmd, shell=True, capture_output=True)
-    
-    if delete:
-        os.remove(job_path)
+    if kwargs.get('resubmit', False):
+        submission_cmd = f'sbatch {job_path}'
+        subp = subprocess.run(submission_cmd, shell=True, capture_output=True)
+        if delete:
+            os.remove(job_path) 
+    else:
+        current_time = strtime()
+        potential_results_path = job_path.split('.')[0] + '_result.json'
+        if os.path.exists(potential_results_path):
+            print(f'[{current_time}] {potential_results_path} already exists, skipping submission')
         
 def get_trigger_time(lightcurve_path):
     '''
